pyproject.toml
[tool.poetry]
name = "swole-bot"
version = "0.1.0"
description = "An AI bot using RAG to help users get swole 💪🏻"
authors = ["Alex <alex@leetsoftware.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.10"
pytest = "8.0.0"
python-dotenv = "1.0.1"
requests = "2.32.3"
tqdm = "4.67.0"
chromadb = "0.5.18"
openai = "1.54.4"
urllib3 = "2.2.3"
transformers = "4.46.2"
torch = "2.5.1"
sentence-transformers = "3.3.0"
pandas = "2.2.3"
backoff = "2.2.1"
llama-index = "0.11.23"
llama-index-core = "0.11.23"
llama-index-vector-stores-chroma = "0.3.0"
llama-index-llms-openai = "0.2.16"
llama-index-embeddings-jinaai = "0.3.1"
llama-index-postprocessor-jinaai-rerank = "0.2.0"
python-fasthtml = "^0.10.0"
markdown = "^3.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
app = "main:main"
